pos/requirements.txt
mysql-connector-python
fastapi
flask

